requirements.txt
streamlit
pandas
plotly
supabase
aiohttp
python-dotenv
openpyxl